third_party/WebKit/LayoutTests/css3/flexbox/justify-content-space-between.html

<!DOCTYPE html>

<style>
.flexbox {
  display: flex;
  flex-direction: column;
  justify-content: space-between;

  min-height: 500px;
}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<body onload="checkLayout('.flexbox')">

<div class="flexbox" data-expected-height="500">
  <div>First item</div>
  <div>Second item</div>
</div>
